student_predictor/requirements.txt
flask
pandas
numpy
scikit-learn
matplotlib
mysql-connector-python
gunicorn
python-dotenv
